Move InMemoryBlobStore to test directory
